refactor(soup): remove unused `soup.rs` module and inline `Soup` into `base.rs`
- Deleted `soup.rs` and adjusted references across modules to use `base::Soup`.
- Simplified conditional patterns with `&& let` syntax where applicable.
- Removed redundant `.into()` calls on `Range` types.
- Streamlined iterator logic in `soup_folding_range`, `hover`, and `validator`.

+++ crates/semantic-tokens/src/soup/value.rs
@@ -142,7 +142,7 @@ fn get_offset(src: &str, pos: tower_lsp_server::ls_types::Position) -> usize {
             line += 1;
             character = 0;
         } else {
-            character += c.len_utf16() as usize;
+            character += c.len_utf16();
         }
     }
     src.len()
